FIX: Handle interpolated values in createUnique
When using an interpolated value as a unique attribute, this method was
saving `undefined` on the `pk`. This commit updates the logic to run the
interpolation before extracting the unique field value from properties.

This commit also adds a test for this case.

NOTE: It isn't clear that `new Expression` is required. Comments above
`prepareProperties` suggest that it is, but manual testing suggests
otherwise.